MAAP-AWS-Meta/main/requirements.txt
boto3==1.36.11
python-dotenv==1.0.1
gradio==5.14.0
filetype==1.2.0
langchain-aws==0.2.12
langchain-mongodb==0.4.0
python-multipart==0.0.20
fastapi==0.115.8
uvicorn==0.34.0
tavily-python==0.5.1
sse-starlette==2.2.1
httpx==0.27.2